[REFACT] LT with BplusTree extensions

+++ include/mementar/core/memGraphs/Branchs/types/SoftPoint.h
@@ -9,8 +9,8 @@ namespace mementar {
 class SoftPoint
 {
 public:
-  //typedef size_t Ttime;
-  typedef float Ttime;
+  typedef size_t Ttime;
+  //typedef float Ttime;
 
   SoftPoint(Ttime t_start, std::experimental::optional<Ttime> t_end = std::experimental::nullopt)
   {
